+++ NEWS.md
@@ -3,7 +3,9 @@
 ## Breaking Changes
 
 ## New Features
-1. Fix @277 - Add function `ts_auto_arima()` boiler plate function.
+1. Fix #277 - Add function `ts_auto_arima()` boiler plate function.
+2. Fix #284 - Add functions `color_blind()` `ts_scale_fill_colorblind()` and
+`ts_scale_color_colorblind()`
 
 ## Minor Fixes and Improvements
 1. Fix #275 - Correct `ts_auto_arima_xgboots()` when `.tune` is FALSE.