+++ NEWS.md
@@ -8,6 +8,7 @@
 `ts_scale_color_colorblind()`
 3. Fix #278 - Add function `ts_auto_smooth_es()`
 4. Fix #279 - Add function `ts_auto_theta()`
+5. Fix #280 - Add function `ts_auto_lm()`
 
 ## Minor Fixes and Improvements
 1. Fix #275 - Correct `ts_auto_arima_xgboots()` when `.tune` is FALSE.