+++ NEWS.md
@@ -1,6 +1,7 @@
 # healthyR.ts (development version)
 
 ## Breaking Changes
+None
 
 ## New Features
 1. Fix #277 - Add function `ts_auto_arima()` boiler plate function.
@@ -9,6 +10,7 @@
 3. Fix #278 - Add function `ts_auto_smooth_es()`
 4. Fix #279 - Add function `ts_auto_theta()`
 5. Fix #280 - Add function `ts_auto_lm()`
+6. Fix #281 - Add function `ts_auto_svm_poly()`
 
 ## Minor Fixes and Improvements
 1. Fix #275 - Correct `ts_auto_arima_xgboots()` when `.tune` is FALSE.