+++ NEWS.md
@@ -1,11 +1,17 @@
 # healthyR.ts (development version)
 
 ## Breaking Changes
+None
 
 ## New Features
 1. Fix #277 - Add function `ts_auto_arima()` boiler plate function.
 2. Fix #284 - Add functions `color_blind()` `ts_scale_fill_colorblind()` and
 `ts_scale_color_colorblind()`
+3. Fix #278 - Add function `ts_auto_smooth_es()`
+4. Fix #279 - Add function `ts_auto_theta()`
+5. Fix #280 - Add function `ts_auto_lm()`
+6. Fix #281 - Add function `ts_auto_svm_poly()`
+7. Fix #282 - Add function `ts_auto_svm_rbf()`
 
 ## Minor Fixes and Improvements
 1. Fix #275 - Correct `ts_auto_arima_xgboots()` when `.tune` is FALSE.